Move systhreads into the runtime (proof of concept).
