+++ Makefile.upstream
@@ -1134,6 +1134,7 @@ runtime_COMMON_C_SOURCES = \
   str \
   sync \
   sys \
+  threads \
   $(TSAN_NATIVE_RUNTIME_C_SOURCES) \
   $(UNIX_OR_WIN32) \
   weak
