+++ runtime/st_win32.h
